requirements.txt
openai
requests
xmltodict
python-dotenv